Backend/requirements.txt
python-dotenv
openai
fastapi
uvicorn
reportlab 